+++ packages/zosfiles/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Zowe z/OS files SDK package will be documented in this file.
 
-## `7.19.0`
+## Recent Changes
 
 - Enhancement: Add data set creation template for COBOL source codes and compiler listings [#1972](https://github.com/zowe/zowe-cli/issues/1972)