+++ packages/zosfiles/CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to the Zowe z/OS files SDK package will be documented in thi
 
 ## `8.0.0-next.202311132045`
 
-- Breaking: First major version bump for V3
+- Major: First major version bump for V3
 
 ## `7.18.9`